Split DumpHeapCommand into a command and service
